+++ NEWS.md
@@ -18,6 +18,7 @@ Version 0.17.2b4 - dev
 - NEW: add `transparency` argument to `LayerTable.add()`
 - NEW: support for transparency `BYLAYER` and `BYBLOCK` for the `drawing` add-on
 - NEW: `Textstyle.make_font()` returns the ezdxf font abstraction
+- NEW: added `dxfattribs` argument to method `Drawing.set_modelspace_vport()`
 - BUGFIX [#574](https://github.com/mozman/ezdxf/issues/574):
   flattening issues in `Path()` and `ConstructionEllipse()` 
 - BUGFIX: `drawing` add-on shows block references in `ACAD_TABLE` at the